+++ server/component/moduleFormsAction/ModuleFormsActionView.php
@@ -255,6 +255,14 @@ class ModuleFormsActionView extends BaseView
                     ))),
                 )),
                 new BaseStyleComponent("descriptionItem", array(
+                    "title" => "Attachments",
+                    "locale" => "",
+                    "id" => "attachments",
+                    "children" => array(new BaseStyleComponent("rawText", array(
+                        "text" => isset($this->action["schedule_info"]['attachments']) ? $this->action["schedule_info"]['attachments'] : ''
+                    ))),
+                )),
+                new BaseStyleComponent("descriptionItem", array(
                     "title" => "Config",
                     "locale" => "",
                     "id" => "config",
@@ -462,6 +470,15 @@ class ModuleFormsActionView extends BaseView
                     "css" => "mb-3",
                     "placeholder" => "@user_name can be used for showing the user",
                 )),
+                new BaseStyleComponent("textarea", array(
+                    "label" => "Attachments",
+                    "type_input" => "json",
+                    "name" => "schedule_info[attachments]",
+                    "id" => "attachments",
+                    "value" => isset($this->action["schedule_info"]['attachments']) ? $this->action["schedule_info"]['attachments'] : '',
+                    "css" => "mb-3",
+                    "placeholder" => "Add attachment files from assets in array",
+                )),   
                 new BaseStyleComponent("ActionConfigBuilder", array(
                     "label" => "Config",
                     "type_input" => "json",